refactor: POST,Userの構造体をmodel.goに切り分けた
